# tools/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Investment Tracker API
  version: 0.1.0
  summary: AI-native portfolio tracker REST API
  description: |
    Read-only aggregator for investment portfolios with an AI-first UX.

    ## Design principles

    - **Spec-first.** This file is the source of truth. Go (huma v2), TypeScript
      (openapi-typescript / openapi-fetch) and Swift (swift-openapi-generator)
      artefacts are generated from this document.
    - **All paths under `/v1/`.** Breaking changes go to `/v2/`.
    - **Cursor pagination** everywhere a list is returned.
    - **Standard error envelope** with a stable `code` and a `request_id`.
    - **Rate limits** are communicated via `X-RateLimit-*` headers on every
      response; `429` is returned when exceeded.
    - **Idempotency.** All mutating endpoints (`POST`, `PATCH`, `DELETE`) accept
      an optional `Idempotency-Key` header. The server stores the response for
      24 hours and replays it for a repeated key.
    - **Money is a string.** Decimal strings (e.g. `"123.4567890000"`) preserve
      precision; JSON numbers would not.
    - **Multi-currency.** Values are always returned in both `base` (USD today)
      and `display` (user preference, converted on the fly).
    - **SSE streaming.** `POST /v1/ai/chat/stream` uses `text/event-stream`.
      A non-stream `POST /v1/ai/chat` counterpart exists for tests and
      retrospectives.
  contact:
    name: Investment Tracker Team
    email: dev@investmenttracker.example
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /me:
    get:
      tags: [Users]
      operationId: getMe
      summary: Get the current user
      responses:
        '200':
          description: Current user
          headers:
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Users]
      operationId: updateMe
      summary: Update the current user's profile
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [Users]
      operationId: deleteMe
      summary: Delete the current user and all associated data (GDPR)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: |
            Deletion accepted. Hard delete is scheduled as an async job;
            from the API's perspective the account is immediately inaccessible.
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /me/paywalls/{trigger}/dismiss:
    parameters:
      - $ref: '#/components/parameters/PaywallTrigger'
    post:
      tags: [Users]
      operationId: dismissPaywall
      summary: Record a soft-paywall dismissal for the current day
      description: |
        Idempotent per (user, trigger, day) — repeated calls within the
        same UTC day are no-ops and return 200. Stored on the shared
        `usage_counters` table with `counter_type='paywall_dismissed_{trigger}'`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Dismissal recorded (or already recorded earlier today)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaywallDismissal' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/undo-deletion:
    post:
      tags: [Users]
      operationId: undoDeletion
      summary: Cancel a pending account deletion (within grace period)
      description: |
        Clears `deletion_scheduled_at`. Only succeeds while the user is
        within the 7-day grace window (see §17.4). After the window expires
        the hard-delete job runs and this endpoint returns `410 GONE`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Deletion cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: Account was not pending deletion
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '410':
          description: Grace period expired — deletion already processed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }

  # ───────────────────────── Security (Clerk-proxied) ────────────
  # All 2FA + session endpoints proxy to Clerk Management API. Clerk is the
  # source of truth — we store no local MFA secret or session row. The
  # endpoints exist so the web/iOS app has a stable surface independent of
  # Clerk SDK changes.
...
  /me/notification-preferences:
    get:
      tags: [Notifications]
      operationId: getNotificationPreferences
      summary: Get per-type channel preferences and digest settings
      responses:
        '200':
          description: Preferences
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Notifications]
      operationId: updateNotificationPreferences
      summary: Update channel preferences / digest settings
      description: |
        Partial update. Omitted `preferences[*]` entries are left unchanged;
        supply only the rows the user actually toggled. Digest settings are
        replaced wholesale when `digest` is present.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationPreferencesUpdate' }
      responses:
        '200':
          description: Updated preferences
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationPreferences' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ──────────────────────────── Exports ──────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer JWT issued by Clerk. Validated against Clerk JWKS on every
        request. `sub` claim is mapped to `users.clerk_user_id`.

  # ─────────────────────── Reusable parameters ───────────────────
  parameters:
...
    PaywallTrigger:
      name: trigger
      in: path
      required: true
      description: Paywall context the client is dismissing (see design brief §13.4).
      schema: { $ref: '#/components/schemas/PaywallTriggerKind' }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Opaque client-generated key (UUID v4 recommended, max 255 chars).
        The server stores the response for 24 hours and replays it for a
        repeated key on the same authenticated user + same endpoint. Keys
        are ignored on `GET` requests.
      schema: { type: string, maxLength: 255 }

  # ─────────────────────── Reusable headers ──────────────────────
  headers:
    RateLimitLimit:
      description: Requests permitted in the current window.
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests remaining in the current window.
      schema: { type: integer }
    RateLimitReset:
      description: Unix epoch seconds when the window resets.
      schema: { type: integer }

  # ─────────────────────── Reusable responses ────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ValidationError:
      description: Request failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
...
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
      description: ISO 4217 currency code.
      examples: ['USD', 'EUR', 'GBP', 'JPY', 'BTC', 'USDT']
...
    SubscriptionTier:
      type: string
      enum: [free, plus, pro]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: |
                Stable machine-readable error code.

                - `TIER_LIMIT_EXCEEDED` — quota exhausted on the current tier
                  (e.g. 5/5 AI messages today). UI shows usage-reset messaging.
                - `FEATURE_LOCKED` — tier does not include this feature at all
                  (e.g. Free/Plus calling `/portfolio/analytics`). UI shows
                  the upgrade/blur-preview treatment from design brief §13.5.
              enum:
                - VALIDATION_ERROR
                - INVALID_REQUEST
                - UNAUTHENTICATED
                - FORBIDDEN
                - TIER_LIMIT_EXCEEDED
                - FEATURE_LOCKED
                - NOT_FOUND
                - CONFLICT
                - DUPLICATE_TRANSACTION
                - IDEMPOTENCY_KEY_CONFLICT
                - IDEMPOTENCY_IN_PROGRESS
                - JURISDICTION_NOT_SUPPORTED
                - QUOTE_NOT_AVAILABLE
                - RATE_LIMIT_EXCEEDED
                - EXTERNAL_SERVICE_ERROR
                - NOT_IMPLEMENTED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable message. Do not parse this.
            details:
              type: object
              additionalProperties: true
              description: Structured details for the error (e.g. failing field).
            request_id:
              type: string
              description: Correlation id; include this in bug reports.
...
    User:
      type: object
      required:
        - id
        - clerk_user_id
        - email
        - display_currency
        - locale
        - subscription_tier
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        clerk_user_id: { type: string }
        email: { type: string, format: email }
        display_currency: { $ref: '#/components/schemas/Currency' }
        locale: { type: string, examples: ['en', 'en-US', 'ru'] }
        subscription_tier: { $ref: '#/components/schemas/SubscriptionTier' }
        stripe_customer_id:
          type: [string, "null"]
        deletion_scheduled_at:
          type: [string, "null"]
          format: date-time
          description: |
            Set when the user initiated account deletion (see §17.4).
            The hard-delete worker runs after a 7-day grace period; the user
            can call `POST /me/undo-deletion` within that window to cancel.
            `null` means the account is active / not pending deletion.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UserUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        display_currency: { $ref: '#/components/schemas/Currency' }
        locale: { type: string }
...
    PaywallTriggerKind:
      type: string
      description: |
        Context key for a soft-paywall modal. Keep in sync with frontend
        constants in packages/ui (see design brief §13.3/§13.4).
      enum:
        - ai_messages_daily
        - insights_weekly
        - connected_accounts
        - pro_feature
    PaywallDismissal:
      type: object
      required: [trigger, dismissed_on, reset_at]
      properties:
        trigger: { $ref: '#/components/schemas/PaywallTriggerKind' }
        dismissed_on:
          type: string
          format: date
          description: UTC day bucket when the dismissal was recorded.
        reset_at:
          type: string
          format: date-time
          description: Next UTC midnight — client MAY re-show the paywall after this instant.
...
    NotificationChannelPreference:
      type: object
      required: [email, push, in_app]
      properties:
        email: { type: boolean }
        push: { type: boolean }
        in_app: { type: boolean }
    NotificationDigestPreferences:
      type: object
      required: [enabled, weekday]
      properties:
        enabled: { type: boolean }
        weekday:
          type: integer
          minimum: 0
          maximum: 6
          description: ISO weekday, 0 = Monday.
        quiet_start:
          type: [string, "null"]
          pattern: '^[0-2][0-9]:[0-5][0-9]$'
          description: Local time HH:mm. iOS-only; mirrored server-side for symmetry.
        quiet_end:
          type: [string, "null"]
          pattern: '^[0-2][0-9]:[0-5][0-9]$'
    NotificationPreferences:
      type: object
      required: [preferences, digest]
      properties:
        preferences:
          type: object
          description: |
            Map keyed by `NotificationType`. Missing keys default to all
            channels enabled.
          additionalProperties:
            $ref: '#/components/schemas/NotificationChannelPreference'
        digest: { $ref: '#/components/schemas/NotificationDigestPreferences' }
    NotificationPreferencesUpdate:
      type: object
      additionalProperties: false
      properties:
        preferences:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NotificationChannelPreference'
        digest: { $ref: '#/components/schemas/NotificationDigestPreferences' }

    # ── FX & prices ───────────────────────────────────────────
...